fix: Unpack blocking websocket broadcasts
The `tar.unpack()` function is a blocking operation, which means it doesn't return until the entire extraction process is complete. Because of this, the async function doesn't yield control back to the Rust runtime until the extraction is finished, which prevents the WebSocket message from being sent.

By moving the extraction process to a separate thread using `tokio::task::spawn_blocking`, you allow the async function to yield control back to the Rust runtime while the extraction is ongoing, which allows the WebSocket message to be sent immediately.