change: Add notifier
We have this weird bug where it wouldn't send a message to a single client during broadcast but it would it was installing but not uninstalling then it would work properly if there is a second client connect and nothing would be blocked. I have an odd feeling this might be something funny going on with the tungstenite crate or how we have implemented the connection handling. This was based off the example server on tokio-tungstenite. The bug is magically fixed by just connecting a second websocket client to the server.