new: Refactor to asynchronous backend (#3)
* change: flavors

* fix: Remove separator

* fix: Luxtorpeda/Boxtron now downloads

* change: clean up

* new: async downloader

* change: more backend stuff

* fix: Unpack blocking websocket broadcasts

The `tar.unpack()` function is a blocking operation, which means it doesn't return until the entire extraction process is complete. Because of this, the async function doesn't yield control back to the Rust runtime until the extraction is finished, which prevents the WebSocket message from being sent.

By moving the extraction process to a separate thread using `tokio::task::spawn_blocking`, you allow the async function to yield control back to the Rust runtime while the extraction is ongoing, which allows the WebSocket message to be sent immediately.

* change: Implement better error handling for github util

* change: cargo fmt

* change: More changes

* change: Renew cache if is 1 day old

* change: Move to async-tungstenite

* change: Fix some bug

* change: Add notifier

We have this weird bug where it wouldn't send a message to a single client during broadcast but it would it was installing but not uninstalling then it would work properly if there is a second client connect and nothing would be blocked. I have an odd feeling this might be something funny going on with the tungstenite crate or how we have implemented the connection handling. This was based off the example server on tokio-tungstenite. The bug is magically fixed by just connecting a second websocket client to the server.

* change: Move from std::sync::Mutex to tokio::sync::Mutex

Turns out this was causing all our blocking issues when locking the app state

* change: Remove RequestState from Task

* change: More logger changes

---------

Co-authored-by: Alberto Hermosillo <themaxhs@gmail.com>